refactor(anyharness): split claude.rs unit tests into claude_tests.rs to stay under line ceiling
